fix guid
